Update Shape/Shape Model.Add Orientation Support For Non-EPD shape. Add SelectedRectF Class To Store Rect With Orientation. Fix TouchHelper Change which lead neo-note nod-epd function work incorrectly issue.

+++ android/kreader/libraries/onyxsdk-scribble/src/main/java/com/onyx/android/sdk/scribble/shape/Shape.java
@@ -76,9 +76,9 @@ public interface Shape {
 
     void resize(final float width, final float height);
 
-    int getOrientation();
+    float getOrientation();
 
-    void setOrientation(int orientation);
+    void setOrientation(float orientation);
 
     void onDown(final TouchPoint normalizedPoint, final TouchPoint screenPoint);
